Moved FiwareEntity to new api sub-package
Moved the FiwareEntity class to a new sub-package called 'api'. Updated all the import statements in other classes to reflect this change. This relocation enhances the organization of the codebase and makes it easier to navigate.
