+++ src/main/java/de/app/fivegla/integration/fiware/model/DevicePosition.java
@@ -1,5 +1,6 @@
 package de.app.fivegla.integration.fiware.model;
 
+import de.app.fivegla.integration.fiware.model.api.FiwareEntity;
 import de.app.fivegla.integration.fiware.model.api.Validatable;
 import de.app.fivegla.integration.fiware.model.internal.Attribute;
 import org.apache.commons.lang3.StringUtils;
